chore(3.5): 升级适配 BBSFS 2.6，编译/构建通过并完成 Mixin 运行时存活审计
- Gradle 依赖切到 BBSFS 2.6（1.20.1/1.20.4），mod_version=3.5.0，fabric.mod.json depends.bbs>=2.6
- 删除/裁剪与 2.6 原生重复或冲突功能（A1/A2/A3/A5），完成 C1 事件迁移
- 迁移 2.6 重命名/重写 API：TrackId 轨道体系、Categories、actions() 操作栏、UIList 拖拽、UIModelRenderer.setDistance、UIScreen.getCurrentMenu、IK 运行时值等
- 快速重接 4 个 Mixin：OrbitFilmCameraControllerFollow、UIClipsDoubleClick、UIReplaysEditor、UIReplayListCategory
- 临时停用 15 个目标已在 2.6 删除、需按 B/C 计划功能级重接的 Mixin（源码保留），保证 jar 可加载
- 新增《Mixin运行时审计与重接清单》记录停用项、恢复方式与后续大项

+++ src/main/java/gbeic/bbsplusplus/utils/AudioDropImporter.java
@@ -147,7 +147,7 @@ public class AudioDropImporter
 
         if (timeline != null && directory.equals(audioFolder))
         {
-            int tick = Math.max(0, timeline.fromGraphX(mouseX));
+            int tick = (int) Math.max(0, timeline.fromGraphX(mouseX));
             int layer = Math.max(0, timeline.fromLayerY(mouseY));
 
             for (String name : imported)
